--- a/.vitepress/theme/index.ts
+++ b/.vitepress/theme/index.ts
@@ -1,8 +1,10 @@
+import DefaultTheme from "vitepress/theme";
 import Layout from "./Layout.vue";
 import * as amplitude from "@amplitude/analytics-browser";
 import "./custom.css";
 
 export default {
+  extends: DefaultTheme,
   Layout,
   async enhanceApp() {
     if (typeof window !== "undefined") {