--- a/ch/code/community.md
+++ b/ch/code/community.md
@@ -1,0 +1,32 @@
+---
+comments: false
+---
+
+# 社区
+
+`Frontend Fundamentals`(FF)与社区一起制定好代码的标准。
+
+目前由 Toss 前端分部维护。
+
+## 专题讨论
+
+查看社区中的精彩讨论。超出《Frontend Fundamentals》 文档中的内容，拓宽你对好代码的思考。
+
+- [专题讨论](https://github.com/toss/frontend-fundamentals/discussions?discussions_q=is%3Aopen+label%3A%22%EC%84%B1%EC%A7%80+%E2%9B%B2%22)
+
+## 探讨代码疑虑
+
+如果你有令人疑惑的代码，可以将其发布到 GitHub 论坛上。
+可以在社区中多角度审查你的代码，并与社区一起探讨好代码的标准。
+
+受到广泛共鸣的案例可直接上传到 Frontend Fundamentals 文档中。贡献方法将会稍后公布。
+
+- [在 GitHub 论坛上发帖](https://github.com/toss/frontend-fundamentals/discussions)
+
+## 为好代码标准添加意见
+
+如果对好代码的标准有意见，或者想提出新的观点，可以参与投票选出你认为更好的代码，并留下自己的意见。与社区沟通，共同构建更加丰富而深入的标准。
+
+这可以成为一个契机，帮助你确立判断两段代码之间哪一段更好的标准。
+
+- [查看 A vs B 上的代码](https://github.com/toss/frontend-fundamentals/discussions/categories/a-vs-b)