## 📝 Key Changes
<!-- Describe the purpose of this PR and the problem it resolves. -->
|**Before**|**After**|
|---|---|
|![image](https://github.com/user-attachments/assets/875c2056-6277-4327-b922-73fdb5a45922)|![image](https://github.com/user-attachments/assets/1ec9b39a-e2e7-4888-bb33-c2fd0c4b679c)|

Browser rendering sequence modifications and operation optimizations have enhanced Performance and Practice scores. These changes specifically impacted the Practice scores of LCP (Largest Contentful Paint) and FCP (First Contentful Paint). The detailed improvements are as follows:

---
## **1. Changes in CSS Loading Strategy**

The frontend-fundamentals previously loaded fonts through CDN CSS files, which acted as an image rendering blocking element. In the new code, we utilized `preconnect`, `fetchpriority="low"`, and `media="none"` to lower CSS loading priority and implement non-blocking processing, resulting in the following benefits:

### **(1) Removal of Render Blocking**

- **Previous Code**: CSS files were considered render-blocking resources by default, delaying image rendering until CSSOM construction was complete.

```jsx
media: "none",
onload: "this.onload=null; this.media='all'",
```

- **After Changes**:
    - Using `media="none"` to exclude CSS from render-blocking elements during initial loading
    - Delaying CSS application timing through the `onload` event to improve initial rendering speed of critical images

### **(2) Resource Priority Optimization**

```jsx
fetchpriority: "low",
```

- By lowering the network request priority of CSS files through `fetchpriority="low"`, critical resources like hero images load first
- As a result, hero images appear on screen faster, reducing FCP and LCP

The translation maintains technical terminology and formatting suitable for a developer's PR, while ensuring clarity in explaining the technical improvements made to the codebase.

---

## **2. Optimization of Hero Image Loading**

Hero images are LCP elements, and are the most important content that should be displayed to users first. The new settings (`loading: eager`, `fetchpriority: high`, `decoding: async`) have optimized image loading and decoding, resulting in performance improvements.

**(1) `loading: eager`**

- By default, browsers load images lazily or when needed.
- `loading: eager` instructs the browser to load hero images immediately, resulting in faster display to users.

**(2) `fetchpriority: high`**

- The browser sets a high network request priority, ensuring that hero images load before other resources.

**(3) `decoding: async`**

- Image decoding is processed asynchronously to prevent blocking the main thread.
- Other rendering tasks (CSSOM construction, DOM updates, etc.) proceed in parallel, resulting in faster overall rendering.

---

## **3. Optimization of Browser Rendering Pipeline**
|Before|After|
|---|---|
|![image](https://github.com/user-attachments/assets/5e3627d7-04e3-4b64-99a9-8d1b4f9b4d10)|![image](https://github.com/user-attachments/assets/3f248063-6b53-4318-9515-ab316e3b73e0)|

The new settings have mitigated bottlenecks in the browser's rendering pipeline:

1. **DOM Construction**: HTML parsing and DOM construction.
2. **CSSOM Construction**: Previously, CSS file loading and parsing were blocked, but now handled in a non-blocking manner.
3. **Render Tree Construction**: After combining DOM and CSSOM, hero images render faster.
4. **Layout and Painting**: Hero images are prioritized as high-priority resources, resulting in faster display on the screen.

---

## **4. Optimization of Network Bottlenecks**

Previously, CSS and images were requested simultaneously, but due to the high priority of CSS, there was a possibility that image requests would be delayed. The new settings have resulted in the following improvements:

- CSS is handled with `fetchpriority="low"`, reducing network bandwidth usage.
- Hero images are processed with `fetchpriority="high"`, resulting in immediate download.
- As a result, network resource competition is reduced, and important content (LCP elements) loads faster.

---

**5. Performance Improvement Summary**

| Change | Reason |
| --- | --- |
| `preconnect` | Preconnect to the CDN to reduce DNS lookup time and improve resource loading speed. |
| `fetchpriority="low"` (CSS) | Lowering the CSS loading priority resulted in hero images loading faster. |
| `media="none"`, `onload` | Removing initial rendering blockage and improving DOM update speed through delayed CSSOM construction. |
| `loading: eager` (이미지) | The hero image loads immediately, resulting in faster display to users. |
| `fetchpriority: high` (이미지) | The network request priority is increased, resulting in faster download of important resources. |
| `decoding: async` (이미지) | Image decoding is processed asynchronously to prevent blocking the main thread and increase parallel operations. |

---

## **Why Practice Score Improved**

![image](https://github.com/user-attachments/assets/15610a61-c9e8-4741-9f17-a8299cfaea9f)

I identified that requesting two CSS files was being recognized as Third-party cookies.

`crossorigin="anonymous"`

However, in the case of requesting two CSS files, no authentication header is required. **`anonymous`** mode does not include cookies/HTTP authentication information, so changing the request mode to anonymous resolves this issue.


## 🖼️ Before and After Comparison
<!-- Attach screenshots or a GIF showing the before and after changes. -->
|**Before**|**After**|
|---|---|
|![image](https://github.com/user-attachments/assets/875c2056-6277-4327-b922-73fdb5a45922)|![image](https://github.com/user-attachments/assets/1ec9b39a-e2e7-4888-bb33-c2fd0c4b679c)|